requirements.txt
pytesseract>=0.3.10
easyocr>=1.7.0
requests>=2.31.0
keyboard>=0.13.5
Pillow>=10.0.0
mss>=9.0.1
opencv-python>=4.8.0
